requirements.txt
maafw
loguru
numpy
strenum
MaaAgentBinary
